+++ NEWS.md
@@ -12,6 +12,7 @@
 * Use SECURE_STORE_ENV, not ENV, for secure store key (MODROLESKC-326)
 * Role with null description becomes invalid (MODROLESKC-308)
 * Improve loadable role capability assignment logic (MODROLESKC-315)
+* Add automatic capability duplicate removal during tenant initialization (MODROLESKC-343)
 * Implement retry logic for calls to keycloak (MODROLESKC-340)
 
 ## Version `v3.0.0` (14.03.2025)